Fix inaccurate ratio of mass to circle radius
Assuming the black circles have uniform density and depth, then the radius varies with the *square root* of the mass, not linearly with the mass. This is because the mass varies linearly with the *area* of the circle.

I multiplied the result by 5 to get circles that were roughly the same size as before (though, necessarily, a smaller range of sizes).

Since I added a new line to calculate the radius on, you will have to update the Gist-slice line numbers in `double-pendulum.haml`. I didn’t do the updates myself because your live site seems to have line numbers for a different version of the code than the `master` branch (which this commit is based on). So you would eventually have to update the line numbers yourself anyway.